TVG_49 Rename database index file
