+++ database/models/UserSearchSubscriptionModel.py
@@ -2,7 +2,7 @@ from sqlalchemy import Column, ForeignKey, Integer
 from sqlalchemy.orm import Mapped, relationship, Session
 from typing import TYPE_CHECKING
 
-from database.database import Base, engine
+from database import Base, engine
 if TYPE_CHECKING:
     from database.models import SearchItem, User
